cloudrun-voice-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
google-auth==2.29.0
requests==2.32.3
